+++ src/lib/adapters/index.ts
@@ -6,7 +6,13 @@ export { superformClient } from './superform.js';
 export { typebox, typeboxClient } from './typebox.js';
 export { valibot, valibotClient } from './valibot.js';
 export { yup, yupClient } from './yup.js';
-export { zod, zodClient, type ZodValidation, type ZodObjectTypes } from './zod.js';
+export {
+	zod,
+	zodClient,
+	type ZodValidation,
+	type ZodObjectTypes,
+	type ZodObjectType
+} from './zod.js';
 export { vine, vineClient } from './vine.js';
 export { schemasafe, schemasafeClient } from './schemasafe.js';
 export { superstruct, superstructClient } from './superstruct.js';