+++ src/lib/adapters/zod.ts
@@ -36,7 +36,7 @@ type ZodObjectUnion<T extends AnyZodObject> = ZodUnion<
 	[ZodValidation<T>, ZodValidation<T>, ...ZodValidation<T>[]]
 >;
 
-type ZodObjectType = ZodType<Record<string, unknown>, ZodTypeDef, Record<string, unknown>>;
+export type ZodObjectType = ZodType<Record<string, unknown>, ZodTypeDef, Record<string, unknown>>;
 
 export type ZodObjectTypes = AnyZodObject | ZodObjectUnion<AnyZodObject> | ZodObjectType;